# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Réservation",
        "description": "API for booking properties and apartments",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/user/bookings": {
            "get": {
                "tags": [
                    "User Bookings"
                ],
                "summary": "Get all user bookings",
                "operationId": "6248872f97b570489c94b73a6b8fa066",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "146f03d075c30736d8c407d5bf42821d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "apartment_id",
                                    "start_date",
                                    "end_date",
                                    "guests_adults"
                                ],
                                "properties": {
                                    "apartment_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-20"
                                    },
                                    "guests_adults": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "guests_children": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/bookings/{booking}": {
            "get": {
                "tags": [
                    "User Bookings"
                ],
                "summary": "Get booking details",
                "operationId": "da09508d1f2270cf0693e0307b6d990f",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Bookings"
                ],
                "summary": "Update a booking",
                "operationId": "d88f499278da5c8d62275fd006024d9c",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-16"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-21"
                                    },
                                    "guests_adults": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 1,
                                        "example": 9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/bookings/{booking}/cancel": {
            "put": {
                "tags": [
                    "User Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "ca8dbe457bb596518067c9679c0af8a8",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "cancelled."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BookingResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "apartment_name": {
                        "type": "string",
                        "example": "Apartment 101"
                    },
                    "apartment_view_link": {
                        "type": "string",
                        "example": "http://example.com/api/v1/apartments/view/1"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-15"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-20"
                    },
                    "adult_guests": {
                        "type": "integer",
                        "example": 2
                    },
                    "children_guests": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "rating": {
                        "type": "integer",
                        "example": 9
                    },
                    "review_comment": {
                        "type": "string",
                        "example": "Great stay!"
                    },
                    "booked_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-01"
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}